Now fourth button toggles a zoom of the preview of camera view to a point where one screen pixel is one recording pixel in order to aid focusing in the field.
For the time being, this means there is no way through the interface
to turn off the program and shut down the pi (the former use of the
fourth button) but it is not clear how often this feature was used
instead of just pulling the power.  If we need it, we can re-add it
(implementing a UI that can allow both this new feature and the old
feature (perhaps using the touch screen? Not sure if that is a good
idea without a lot of testing in field conditions (concerns about
gloves, cold fingers, temp affecting touch screen functioning...)))