Move dencam_setup.txt to a folder for notes
PBI was using these notes when they were configuring Raspberry Pis to
run DenCam.  We should keep them in the repo for their reference but
should also try to flesh out the README to cover all the setup
instructions necessary as the README is, of course, where folks will
typically look for setup instructions and it would be better to have
more explanation associated with each of the steps.