Implement state list
Added a state list variable to the yaml file, now instead of initializing
the number of states with a hardcoded number we can use the list's length.
In parts where we need to check if state.value is equal to a certain number,
we can now check to see if state.value corresponds to the index of the page
name.