Code fix
Removed STATE_LIST from yaml and is now a variable created in dencam.py,
which is then passed to the classes that need it. Fixed the numbering
issue so there is no more +1/-1 when using the index of the list. Fixed
naming convention of the variable, is state_list in arguments, STATE_LIST
otherwise.