Handles camera connectivity alert
Dencam will simply crash if the camera is not detected on startup
and will not rerun. By adding an exception and a screen to display
the issue when the camera is not detected, users can resolve the
issue and then the program will continue to run as normal, once
the exception is no longer thrown. error_handling.py takes care
of the screen alert and is called in the try except block in main
and the screen is destroyed if the try is successful.

Resolves #66