file not found for solar screen more verbose
The solar screen displayed "file not found" when the solar log
was not available to the script. this means that the sunsaver_log.py
script hadn't been run, which needs to happen as a cronjob as per
the readme. it does not mean that solar is not connected to the pi,
just that there is no solar log file available. the solar might be
plugged in, just not writing to that log file yet. the solution to
this screen is to run sunsaver_log.py, and then if the solar
is truly not connected then there will be a different error that
shows after that- indicated by all the fields in the log file
being marked as "NA"