beginnings of trying to make cyclops.py run dencam with ptz
going to use aim_ptz.py now to not be the camera test, that will
be ptz_control_test.py. this is because i will have to change how
that script is formatted, more class-ish, so that i can call it in
the new "recorder.py" that will be replaced for now by
"cyclops.py" and which will need to call that class stuff for
initializing ptz camera view and controls. bit messy right now
and definitely doesn't work, but committing to save my progress
in case the pi does something weird and loses it.
essentially hoping to get cyclops.py with cyclops_config.yaml
to run dencam, but with the recorder class replaced by one that
is ptx friendly and calls the aim_ptz class to do its camera
initialization as recorder used to handle it for picamera, but
this one is more complicated so id rather keep it encapsulated in
aim_ptz. ideally, it will look like dencam but instead of toggling
zoom, youd be toggling the ps4 controller aspect, but stopping
using the ps4 to control has yet to be implemented. probably need
to write that first if that's how i want it to work or get rid of
that altogether for now