make the control at least start to work
the control at least controls a little bit, not perfectly and not
on command. need to troubleshoot more about how the ps4 controller
stuff is called and created. and also need to still fix the issue
where the display doesn't get created on the second go around
need to also make the output not output every .0002 seconds
what the position of the camera is. need to also make the control
of the ptz toggle with button press like how picamera zooms