Refactor prep_fonts() into method of Controller
Makes the font preparation function into a private method of the
Controller class and also makes the fonts dict itself an attribute of
the Controller class.

Arguably, in a true MVC implementation, fonts wouldn't belong in the
Controller but we've stuffed a lot of other things in there and it
feels like what this class actually is is something that deserves to
own the fonts.