make ptz display properly display upon each go through of the ui
so before, you could only initialize the ptz viewing screen on the
first go through of the dencam screens, now it will properly
delete and then show the screen every time you loop through dencam
screens. only issue is that it takes an extra button press to exit
out of the preview, which could be worked on, but at the very least
it properly loops through properly each time. ideally it would
take 1-2 button presses, currently its 3 to get to the next screen