loop through pages seamlessly without error
previous commit had a bug wherein there would be a segmentation
fault occassionally on the second round through the screens, i
believe it was due to erroneous button handling in the camview
class and not initializing the screen in the proper order, causing
it to be slow to bring the display up. now it quickly loops
through the screens. one improvement can be the extra button
press required to get back to the main page, could remove the
"blank page" as that was only necessary for the picamera
implementation. removing the blank screen would cause the same
user feel as minidencam, as a blank screen is introduced after the
ptz display is destroyed