add worded button labels for the pages
before it was button icons, now there are worded button labels
for the buttons, will need to be merge corrected with dev because
the version addition will disrupt buttons, but it works well with
this version. the camera preview page has a label for the second
button but it has to be at the bottom of the screen.