Refactor: move picamera-specific bits out of Recorder
In prep for adding code to handle a different kind of camera (i.e. a
ptz network surveillance camera for cyclops), refactoring out
picamera-specific bits of `BaseRecorder` and into a derived class (via
`Recorder`).  The derived class is in a separate module so that the
picamera package is only required in that module.  The idea is that we
will make a similar derived class for the ptz camera, and
`BaseRecorder` and `Recorder` (and thus ideally everything in
`recorder.py`) will be truly agnostic to the difference systems.

An assumption is that we will wrap the ptz camera control code in
methods that match those of `picamera.PiCamera` so that every place in
`recorder.py` that `self.camera` uses those methods it will all work.

Ideally, we will also be able to do a similar thing with changes
necessary for picamera2 i.e. make class derived from `Recorder` in its
own module that implements everything as it needs to be for using
picamera2.  That is part of the reason we didn't use an alternate
solution: making a camera object an argument to `BaseRecorder`
constructor and initializing the camera in the top-level application
module.