fix hardcoded pi user in mppt.py
user must add their file path to dencam codebase in the config
file before running dencam so that the mppt.py and sunsaver_log
knows where to get the solar.csv. the config parser needs to be
in mppt.py because sunsaver_log is ran standalone on a loop
underneath dencam, and so mppt.py needs to contain the full
capability of getting that file path without being passed it by
dencam itself.

Resolves #80