make recording and picamera2 display toggle
these changes allow for recording to take place and the camera
preview to be toggled throughout looping through the dencam pages.
picamera2 is picky when it comes to if it will record and preview
simultaneously, and the solution is to set preview and record
to different streams from the camera, the preview getting the
lower quality one the the recording getting the higher quality one.
Changes still need to be made to make the preview the correct size
on the screen because right now it's too big, but it properly stops
and starts and recording is able to occur on the main stream without
distrupting the preview