Instruct pylint to ignore perceived self-import
Because this application program module has the same name as the
package (perhaps not best practice), pylint gets worried about self
import. We could rename this module but that would mean combing
through and changing all references to it in the README and perhaps
other gotchas so for now punting on this and telling pylint to ignore
its worries.

Maybe there is a third option that allows us to keep this name and not
excite pylint.  That is left as an exercise for the reader.