modularizing picamera2 functionality outside recorder.py
also adding in the ability to define which camera library
will be used in the config file, this way error handling for
if the camera isn't connected can be achieved the same way as
before in the dencam.py script. this should definitely be made
more efficient but for now it works without too much error
handling. now i need to ensure all the config values
for the picamera one can be converted to picamera2