adding classes for differing gui pixels depending on os
there were 2 different pixel values for gui placements hardcoded
depending on if you were running buster (fenrir) or bookworm
(lesehest and mimir). Now, there is a module with 2 classes and
gui, when it creates the controller object, will figure out
if its running buster or bookworm and load in the appropriate
class of dictionary with the corresponding pixel values. Each
screen class in gui.py had to be changed to accept these
dictionary values for placing the objects. All that's left is to
clean it up (pylint, flake8)

Resolves #107