Implement required "zoom" functionality
The DenCam software exposes in the UI the ability to "zoom" in on the
video display page.  Now `MimirCamera` implements the expected
interface/functionality for this "zoom" parameter.  I keeping putting
the word zoom in parentheses because we should probably rename it as
it is merely a digital zoom that is only in the display whose purpose
is to aid in focusing.  There's been confusion about that that might
be alleviated by renaming the variable (in this module and maybe
higher up ideally) especially since in Mimir we have programmatic
control over optical zoom.