Rename this class
(we have too many things named "controller" in this project what with
UI controllers etc. The term is totally overloaded.  I can see the
argument that a PS4 controller is actually more than a joystick and in
fact has itself two joysticks but it is the best term I can think of
quickly and certainly creates much less confusion than having another
"controller" in the mix)