make lesehest toggle zoom button work
currently toggles zoom in and out, had to change the implementation
of toggle zoom inside recorder.py and rewrite the function. what
it used to do in recorder.py will need to go in recorder_picamera.py
because i had to have it just call the rewritten class in
recorder_picamera2.py because picamera and picamera2 do
zooming differently. also the user experience is a little different
than before, with this one you see it zoom smoothly in and it takes
about 1-2 seconds, but i think it makes it clearer what you're
zooming in on, so i think its an improvement, but could also be
changed to go back to just immediately zooming in and out.