Refactor `zoom_toggle` method
In both of the zooming out implementation and the zooming in
implementation there were a lot of the same function calls and it made
sense to refactor this so that it is clearer what is common and what
is different about the actions taken in the two states.

Also, gave a magic number a name.  Decided for now not to make this a
module constant but just variable local to the method.