fix background box size and lint
the background black box behind the text was a bit off, now
it's good. What this does is add the timestamp to both the preview
and the recording. The preview will maintain the timestamp even
while zooming in and out. Though it does seem to make the zoom process
SLIGHTLY slower. It would need to be completely changed in implementation
if we wanted to not display the timestamp on the preview as well
due to changes between picamera and picamera2

Resolves #130