fix gui widget placement for network page
now is in correct orientation for flipping 180 degrees. I made the
decision to move the page label to the bottom, because with the 180
flip, the bottom button is higher than the previous ui assumes.
This limits the space that the actual information on the page has,
if you still have the page label on the top. This way, the
information on the page has more space that is not disrupted by
the button labels. The words will automatically center themselves.
If we wanted to have each line in the information to be each truly
centered and not right-justified in the center, networking.py
would need to be modified.