first commit with the basic changes
adds the functions from the picamera2 focus score utility to the
3rd button on the preview screen. major issue, once you toggle
the focus score once i can't turn it off with this current
implementation. it also freezes essentially on that focus score
screen, and the laplacian values are super wack when you are
focused while doing the focus score. It also thinks the timestamp
is something in the frame that it has super high focus score for.
Otherwise it does toggle on when that third button is pressed, and
works. Just needs to be toggled off, and ignore the timestamp
values and be able to adjust when it's focused- specifically
achieved in that order