Make toggle focus only flipflop boolean
Ie our plan is to use this boolean in the callback that does overlay
drawing (currently is only drawing the timestamp) so the toggle method
only needs to flip to boolean.

Also: in `buttons.py` removed the wait before this button is active as
such waiting is not desired for this button/function.

Also: renamed the toggle focus display method to be a bit more
explanatory (as isn't really affecting focus just bringing up the
display).