Implement non-blocking focus score display toggling
Uses same callback function that renders timestamps.  Since we don't
really want timestamps when using the focus score tool, hides
timestamp when focus score display is on.

Third button controls the focus score display toggle.

In the process of this implementation, moved the function that
calculates the focus score out of the class it had been embedded in
and adapted the method that calculates that score for each element in
the grid and renders that information to the screen to jive with the
callback-based mechanics.