change that should work but has issues
this change allocates the proper stream types to their
recording/previewing, (the raw stream must also be there and
can't be smaller than the biggest stream). this should work,
but it causes an error where it says it cannot allocate memory
picamera2 doc (pg 58) and online forums mention to increase
cma memory allocation, and this appeared to completely work from
the viewpoint of the log messages- but the screen was totally blank
this is likley because we don't use the dtoverlay that is required
to implement picamera2's solution to the allocation errors, and seemed
to conflict with our pitft screen dtoverlay parameter. i tried
to increase cma from elsewhere, but it just gave me a different issue.
essentially, this should work- but perhaps you can decrease frame buffers
for the streams and it could decrease memory needed and will get rid of the
issue, i just didn't have time tonight. but basically, increasing cma
made this solution worked, but it messes with our screen overlay so
we can't copy picamera2 doc solution as is (unless my screen got buggy
for another reason)