Fix (quick fix) some issues with zoom control
There have been two issues that now seem to have the same root cause:
1) the system zooming back out from wherever it was zoomed to when a
pan or tilt move is done and 2) not being able to zoom beyond one zoom
level when incrementally zooming. The problem is believed to come from
`ptzipcam.ptz_camera.get_position()` erroneously returning zero for
zoom (maybe only some of the time? because otherwise why is the
proportional pan working since it takes this value into account).  So,
this fix has the `Joystick` class actively track the zoom level
itself.  This is a quick and dirty fix at this stage (necessary to
capture now because of need to demo), mostly because it introduces a
different bug: the zoom needs to be set to a known value whenever the
`Joystick` class is constructed which is currently every time the user
returns to the "Preview" page so, while the system will retain pan and
tilt position when that happens, the camera will always then zoom out,
which is undesirable behaviour.