Make pan/tilt button actions reflect camera orientation
What is left/right and up/down depends on whether camera is dome up or
down.  For current design, camera is always down while code assumed up
as it was written before that change. Rather than simply reverse
movements, the code now toggles appropriate movement directions based
on a boolean module constant.

Perhaps at some point it would be worthwhile to get this boolean value
from the config file but, at this stage, systems can be assumed to have
camera dome down so that's not a priority.