Make quick fix from d524ca3 better/different
The quick fix implemented in d524ca3 had the undesirable effect that
while it did get rid of the unexpected zooming it out, it introduced
another problem: every time you cycled away from the preview page and
back or restarted the program, you'd lose your zoom state.  That is
really quite undesirable for the application. Some work was done to
expand the hack by having the `zoom_state` stored at a higher level of
the class stack in a class that persists for the run-time of the
program but even that wouldn't survive a power cycling or re-run. So,
we have gone back to relying on the PTZ camera to hold zoom state and
give it to us.  But because we know it is flakey with that below some
minimum zoom threshold (i.e. it gives us 0.0 even if it is at eg 0.7;
there's definitely a chance that this bug is not with the camera but
somewhere in the stack of libraries we are using to talk to the
camera), our program now forces the zoom level to stay above that
threshold.